pyproject.toml
[project]
name = "django-startapp-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "django>=5.1.5",
    "django-celery-results>=2.5.1",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=24.3",
    "django-redis>=5.4.0",
    "django-silk>=5.3.2",
    "django-unfold>=0.52.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "markdown>=3.7",
]

[dependency-groups]
dev = [
    "django-extensions>=3.2.3",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.4",
]
